Improve scripts
Run test-all so linting etc. is also run during ci tests.
Just search for bash scripts in the script folder to speed things up a bit, this also simplifies the file check.
Ensure scripts are being run from the correct location as some scripts required the directory to be changed as they expect to be in a specific folder.